// src/api/generated/schemas/getKeywords200DataItemVersionsItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */

import type { GetKeywords200DataItemVersionsItemCategory } from "./getKeywords200DataItemVersionsItemCategory";
import type { GetKeywords200DataItemVersionsItemCategoryId } from "./getKeywords200DataItemVersionsItemCategoryId";
import type { GetKeywords200DataItemVersionsItemCreatedAt } from "./getKeywords200DataItemVersionsItemCreatedAt";
import type { GetKeywords200DataItemVersionsItemCreatedById } from "./getKeywords200DataItemVersionsItemCreatedById";
import type { GetKeywords200DataItemVersionsItemDescription } from "./getKeywords200DataItemVersionsItemDescription";
import type { GetKeywords200DataItemVersionsItemEndingChapter } from "./getKeywords200DataItemVersionsItemEndingChapter";
import type { GetKeywords200DataItemVersionsItemImage } from "./getKeywords200DataItemVersionsItemImage";
import type { GetKeywords200DataItemVersionsItemImageId } from "./getKeywords200DataItemVersionsItemImageId";
import type { GetKeywords200DataItemVersionsItemNature } from "./getKeywords200DataItemVersionsItemNature";
import type { GetKeywords200DataItemVersionsItemNatureId } from "./getKeywords200DataItemVersionsItemNatureId";
import type { GetKeywords200DataItemVersionsItemStartingChapter } from "./getKeywords200DataItemVersionsItemStartingChapter";
import type { GetKeywords200DataItemVersionsItemUpdatedAt } from "./getKeywords200DataItemVersionsItemUpdatedAt";

export type GetKeywords200DataItemVersionsItem = {
	id: string;
	description: GetKeywords200DataItemVersionsItemDescription;
	startingChapter: GetKeywords200DataItemVersionsItemStartingChapter;
	endingChapter: GetKeywords200DataItemVersionsItemEndingChapter;
	categoryId: GetKeywords200DataItemVersionsItemCategoryId;
	natureId: GetKeywords200DataItemVersionsItemNatureId;
	imageId: GetKeywords200DataItemVersionsItemImageId;
	keywordId: string;
	createdById: GetKeywords200DataItemVersionsItemCreatedById;
	createdAt: GetKeywords200DataItemVersionsItemCreatedAt;
	updatedAt: GetKeywords200DataItemVersionsItemUpdatedAt;
	/** @nullable */
	category: GetKeywords200DataItemVersionsItemCategory;
	/** @nullable */
	nature: GetKeywords200DataItemVersionsItemNature;
	/** @nullable */
	image: GetKeywords200DataItemVersionsItemImage;
};